Changed get_db_status
It's a little ambiguous as to what 'database is not active' means. I've made a design assumption that 'not active' means that the database doesn't exist (or can't be found for some reason by python)

There are probably more robust ways for checking the status, but unfortunately I'm not that experienced working with sqlite3 (yet..)!